backend/requirements.txt
flask
flask-cors
PyJWT
mysql-connector-python
requests
opencv-python
numpy
tensorflow
Pillow
sqlite3